Added File upload and download Generate invoices

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const { validationResult } = require('express-validator/check');
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 
 exports.getAddProducts = (req, res, next) => {
 	res.render('admin/edit-product', {
@@ -13,12 +14,11 @@ exports.getAddProducts = (req, res, next) => {
 
 exports.postAddProducts = (req, res, next) => {
 	const title = req.body.title;
-	const imageUrl = req.body.imageUrl;
+	const image = req.file;
 	const price = req.body.price;
 	const description = req.body.description;
 	const error = validationResult(req);
-	if (!error.isEmpty()) {
-		console.log(error);
+	if (!image) {
 		return res.status(422).render('admin/edit-product', {
 			pageTitle: 'Add product',
 			path: '/admin/edit-product',
@@ -27,7 +27,23 @@ exports.postAddProducts = (req, res, next) => {
 			product: {
 				title: title,
 				price: price,
-				imageUrl: imageUrl,
+				description: description,
+			},
+			errorMsg: 'Attached file is not an image',
+		});
+	}
+	const imageUrl = image.path;
+
+	if (!error.isEmpty()) {
+		fileHelper.deleteFile(imageUrl);
+		return res.status(422).render('admin/edit-product', {
+			pageTitle: 'Add product',
+			path: '/admin/edit-product',
+			editMode: false,
+			hasError: true,
+			product: {
+				title: title,
+				price: price,
 				description: description,
 			},
 			errorMsg: error.array()[0].msg,
@@ -43,11 +59,12 @@ exports.postAddProducts = (req, res, next) => {
 	product
 		.save()
 		.then((result) => {
-			console.log('Product Created');
 			res.redirect('/admin/products');
 		})
 		.catch((err) => {
-			console.log(err);
+			const error = new Error(err);
+			error.httpStatusCode = 500;
+			return next(error);
 		});
 };
 
@@ -64,7 +81,9 @@ exports.getProducts = (req, res, next) => {
 			});
 		})
 		.catch((err) => {
-			console.log(err);
+			const error = new Error(err);
+			error.httpStatusCode = 500;
+			return next(error);
 		});
 };
 
@@ -90,14 +109,16 @@ exports.getEditProducts = (req, res, next) => {
 			});
 		})
 		.catch((err) => {
-			console.log(err);
+			const error = new Error(err);
+			error.httpStatusCode = 500;
+			return next(error);
 		});
 };
 
 exports.postEditProduct = (req, res, next) => {
 	const id = req.body.productId;
 	const updatedTitle = req.body.title;
-	const updatedImageUrl = req.body.imageUrl;
+	const image = req.file;
 	const updatedPrice = req.body.price;
 	const updatedDescription = req.body.description;
 	const error = validationResult(req);
@@ -110,7 +131,6 @@ exports.postEditProduct = (req, res, next) => {
 			product: {
 				title: updatedTitle,
 				price: updatedPrice,
-				imageUrl: updatedImageUrl,
 				description: updatedDescription,
 				_id: id,
 			},
@@ -123,7 +143,10 @@ exports.postEditProduct = (req, res, next) => {
 				return res.redirect('/');
 			}
 			product.title = updatedTitle;
-			product.imageUrl = updatedImageUrl;
+			if (image) {
+				fileHelper.deleteFile(product.imageUrl);
+				product.imageUrl = image.path;
+			}
 			product.description = updatedDescription;
 			product.price = updatedPrice;
 			return product.save().then((result) => {
@@ -132,18 +155,29 @@ exports.postEditProduct = (req, res, next) => {
 			});
 		})
 		.catch((err) => {
-			console.log(err);
+			const error = new Error(err);
+			error.httpStatusCode = 500;
+			return next(error);
 		});
 };
 
 exports.postDeleteProduct = (req, res, next) => {
 	const id = req.body.productId;
-	Product.deleteOne({ _id: id, userId: req.user._id })
+	Product.findById(id)
+		.then((prod) => {
+			if (!prod) {
+				return new Error('Product not found.');
+			}
+			fileHelper.deleteFile(prod.imageUrl);
+			return Product.deleteOne({ _id: id, userId: req.user._id });
+		})
 		.then((result) => {
 			console.log('Product Deleted');
 			res.redirect('/admin/products');
 		})
 		.catch((err) => {
-			console.log(err);
+			const error = new Error(err);
+			error.httpStatusCode = 500;
+			return next(error);
 		});
 };
